stage削除時に属するtaskがpendingに移動するようにした

+++ stages/views.py
@@ -43,7 +43,6 @@ class StageUpdateView(View):
         if form.is_valid:
             stage = form.save(commit=False)
             stage.plan = get_object_or_404(Plan, pk=plan_pk)
-            stage.order = Stage.objects.count() + 1
             stage.save()
             return redirect("plans:show", plan_pk=plan_pk)
         context = {"form": form}
@@ -51,11 +50,20 @@ class StageUpdateView(View):
 
 
 class StageDeleteView(View):
+    @transaction.atomic
     def post(self, request, plan_pk, stage_pk):
         plan = get_object_or_404(Plan, pk=plan_pk)
-        if request.user != plan.owner:
-            return HttpResponseForbidden("このステージを削除することは禁止されています。")
         stage = get_object_or_404(Stage, pk=stage_pk)
+        # plan非所有者による削除とpending, doneステージの削除を禁止
+        if request.user != plan.owner or stage.order < 0:
+            return HttpResponseForbidden("このステージを削除することは禁止されています。")
+        tasks = stage.task_set.all()
+        for task in tasks:
+            pending = plan.stage_set.get(order=-2)
+            task.stage = pending
+            print(task.name, task.stage)
+            task.order = pending.task_set.count() + 1
+            task.save()
         stage.delete()
         return redirect("plans:show", plan_pk=plan_pk)
